+++ SharePointFramework/PortfolioWebParts/src/components/BenefitsOverview/BenefitMeasurementAchievement/index.tsx
@@ -1,32 +1,19 @@
-import * as React from 'react'
-import { BenefitMeasurement } from 'models'
 import { Icon } from 'office-ui-fabric-react/lib/Icon'
-import * as objectGet from 'object-get'
+import React from 'react'
+import { IBenefitMeasurementAchievementProps } from './types'
 
-export interface IBenefitMeasurementAchievementProps extends React.HTMLProps<HTMLSpanElement> {
-  measurement: BenefitMeasurement
-}
-
-export default class BenefitMeasurementAchievement extends React.PureComponent<
-  IBenefitMeasurementAchievementProps
-> {
-  constructor(props: IBenefitMeasurementAchievementProps) {
-    super(props)
-  }
-
-  public render(): React.ReactElement<IBenefitMeasurementAchievementProps> {
-    const colValue = objectGet(this.props.measurement, 'achievementDisplay')
-    const trendIconProps = objectGet(this.props.measurement, 'trendIconProps')
-    if (colValue) {
-      return (
-        <span>
-          <span style={{ display: 'inline-block', width: 20 }}>
-            {trendIconProps && <Icon {...trendIconProps} />}
-          </span>
-          <span>{colValue}</span>
+export const BenefitMeasurementAchievement = ({
+  measurement
+}: IBenefitMeasurementAchievementProps) => {
+  if (measurement.AchievementDisplay) {
+    return (
+      <span>
+        <span style={{ display: 'inline-block', width: 20 }}>
+          {measurement.TrendIconProps && <Icon {...measurement.TrendIconProps} />}
         </span>
-      )
-    }
-    return null
+        <span>{measurement.AchievementDisplay}</span>
+      </span>
+    )
   }
+  return null
 }