+++ SharePointFramework/PortfolioWebParts/src/components/LatestProjects/index.tsx
@@ -1,23 +1,22 @@
 import { DisplayMode } from '@microsoft/sp-core-library'
 import { SortDirection } from '@pnp/sp'
 import { WebPartTitle } from '@pnp/spfx-controls-react/lib/WebPartTitle'
-import { formatDate } from 'pp365-shared/lib/helpers'
 import { MessageBar } from 'office-ui-fabric-react/lib/MessageBar'
 import { Spinner, SpinnerType } from 'office-ui-fabric-react/lib/Spinner'
 import * as PortfolioWebPartsStrings from 'PortfolioWebPartsStrings'
-import * as React from 'react'
-import { ILatestProjectsProps } from './ILatestProjectsProps'
-import { ILatestProjectsState } from './ILatestProjectsState'
+import { formatDate } from 'pp365-shared/lib/helpers'
+import React, { Component } from 'react'
 import styles from './LatestProjects.module.scss'
+import { ILatestProjectsProps, ILatestProjectsState } from './types'
 
 /**
  * @component LatestProjects
- * @extends React.Component
+ * @extends Component
  */
-export class LatestProjects extends React.Component<ILatestProjectsProps, ILatestProjectsState> {
+export class LatestProjects extends Component<ILatestProjectsProps, ILatestProjectsState> {
   constructor(props: ILatestProjectsProps) {
     super(props)
-    this.state = { isLoading: true, projects: [] }
+    this.state = { loading: true, projects: [] }
   }
 
   public async componentDidMount() {
@@ -27,9 +26,9 @@ export class LatestProjects extends React.Component<ILatestProjectsProps, ILates
         'Created',
         SortDirection.Descending
       )
-      this.setState({ projects, isLoading: false })
+      this.setState({ projects, loading: false })
     } catch (error) {
-      this.setState({ projects: [], isLoading: false })
+      this.setState({ projects: [], loading: false })
     }
   }
 
@@ -38,14 +37,14 @@ export class LatestProjects extends React.Component<ILatestProjectsProps, ILates
    */
   public render(): React.ReactElement<ILatestProjectsProps> {
     return (
-      <div className={styles.latestProjects}>
+      <div className={styles.root}>
         <WebPartTitle
           displayMode={DisplayMode.Read}
           title={this.props.title}
           updateProperty={undefined}
         />
         <div className={styles.container}>
-          {this.state.isLoading ? (
+          {this.state.loading ? (
             <Spinner label={this.props.loadingText} type={SpinnerType.large} />
           ) : (
             this._renderProjectList()