+++ SharePointFramework/PortfolioWebParts/src/components/LatestProjects/types.ts
@@ -1,0 +1,30 @@
+import { IBaseComponentProps } from '../types'
+
+export interface ILatestProjectsProps extends IBaseComponentProps {
+  /**
+   * Loading text
+   */
+  loadingText: string
+
+  /**
+   * Empty message
+   */
+  emptyMessage: string
+
+  /**
+   * Number of items to show
+   */
+  rowLimit: number
+}
+
+export interface ILatestProjectsState {
+  /**
+   * Whether the component is loading
+   */
+  loading: boolean
+
+  /**
+   * Projects
+   */
+  projects: any[]
+}