+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/RenderItemColumn/IRenderItemColumnProps.ts
@@ -1,6 +1,7 @@
 import { ProjectColumn } from 'pp365-shared/lib/models/ProjectColumn'
 
-export interface IRenderItemColumnProps {
-  column: ProjectColumn
-  colValue: string
+export interface IRenderItemColumnProps extends React.HTMLAttributes<HTMLDivElement> {
+  column?: ProjectColumn
+  columnValue: string
+  valueSeparator?: string
 }