+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/RenderItemColumn/TagsColumn/index.tsx
@@ -1,16 +1,21 @@
-import * as React from 'react'
+import { stringIsNullOrEmpty } from '@pnp/common'
+import React, { FunctionComponent } from 'react'
 import { IRenderItemColumnProps } from '../IRenderItemColumnProps'
 import { Tag } from './Tag'
+import styles from './TagsColumn.module.scss'
 
-export class TagsColumn extends React.Component<IRenderItemColumnProps> {
-  public render(): React.ReactElement<IRenderItemColumnProps> {
-    const tags: string[] = this.props.colValue.split(';')
-    return (
-      <span>
-        {tags.map((text, idx) => (
-          <Tag key={idx} text={text} />
-        ))}
-      </span>
-    )
-  }
+export const TagsColumn: FunctionComponent<IRenderItemColumnProps> = ({
+  columnValue,
+  valueSeparator = ';',
+  style
+}: IRenderItemColumnProps) => {
+  if (!columnValue) return null
+  const tags: string[] = columnValue.split(valueSeparator).filter((t) => !stringIsNullOrEmpty(t))
+  return (
+    <div className={styles.root} style={style}>
+      {tags.map((text, idx) => (
+        <Tag key={idx} text={text} />
+      ))}
+    </div>
+  )
 }