+++ SharePointFramework/PortfolioWebParts/src/models/BenefitMeasurementIndicator.ts
@@ -2,15 +2,15 @@ import { IBenefitsSearchResult } from 'interfaces'
 import { BenefitBase, Benefit, BenefitMeasurement } from './'
 
 export class BenefitMeasurementIndicator extends BenefitBase {
-  public indicator: string
-  public startValue: number
-  public desiredValue: number
-  public startValueDisplay: string
-  public desiredValueDisplay: string
-  public unit: string
-  public benefitId: number
-  public measurements: BenefitMeasurement[]
-  public benefit: Benefit
+  public Indicator: string
+  public StartValue: number
+  public DesiredValue: number
+  public StartValueDisplay: string
+  public DesiredValueDisplay: string
+  public Unit: string
+  public BenefitItemId: number
+  public Measurements: BenefitMeasurement[]
+  public Benefit: Benefit
 
   /**
    * Creates a new instance of BenefitMeasurementIndicator
@@ -18,23 +18,23 @@ export class BenefitMeasurementIndicator extends BenefitBase {
    * @param {IBenefitsSearchResult} result Search result
    * @param {number} fractionDigits Fraction digits for valueDisplay
    */
-  constructor(result: IBenefitsSearchResult, fractionDigits = 2) {
+  constructor(result: IBenefitsSearchResult, fractionDigits: number = 2) {
     super(result)
-    this.indicator = result.GtMeasureIndicatorOWSTEXT
-    this.startValue = !isNaN(parseFloat(result.GtStartValueOWSNMBR))
+    this.Indicator = result.GtMeasureIndicatorOWSTEXT
+    this.StartValue = !isNaN(parseFloat(result.GtStartValueOWSNMBR))
       ? parseFloat(result.GtStartValueOWSNMBR)
       : null
-    if (this.startValue !== null) {
-      this.startValueDisplay = this.startValue.toFixed(fractionDigits)
+    if (this.StartValue !== null) {
+      this.StartValueDisplay = this.StartValue.toFixed(fractionDigits)
     }
-    this.desiredValue = !isNaN(parseFloat(result.GtDesiredValueOWSNMBR))
+    this.DesiredValue = !isNaN(parseFloat(result.GtDesiredValueOWSNMBR))
       ? parseFloat(result.GtDesiredValueOWSNMBR)
       : null
-    if (this.desiredValue !== null) {
-      this.desiredValueDisplay = this.desiredValue.toFixed(fractionDigits)
+    if (this.DesiredValue !== null) {
+      this.DesiredValueDisplay = this.DesiredValue.toFixed(fractionDigits)
     }
-    this.unit = result.GtMeasurementUnitOWSCHCS
-    this.benefitId = parseInt(result.GtGainLookupId, 10)
+    this.Unit = result.GtMeasurementUnitOWSCHCS
+    this.BenefitItemId = parseInt(result.GtGainLookupId, 10)
   }
 
   /**
@@ -43,10 +43,10 @@ export class BenefitMeasurementIndicator extends BenefitBase {
    * @param {BenefitMeasurement[]} measurements Measurements
    */
   public setMeasurements(measurements: BenefitMeasurement[]): BenefitMeasurementIndicator {
-    measurements = measurements.filter((m) => m.indicatorId === this.id && m.siteId === this.siteId)
+    measurements = measurements.filter((m) => m.IndicatorId === this.Id && m.SiteId === this.SiteId)
     measurements = measurements.map((m) => m.calculcateAchievement(this))
     measurements = measurements.map((m, i) => m.setTrendIconProps(measurements[i + 1]))
-    this.measurements = measurements
+    this.Measurements = measurements
     return this
   }
 
@@ -56,7 +56,9 @@ export class BenefitMeasurementIndicator extends BenefitBase {
    * @param {Benefit[]} benefits Benefits
    */
   public setBenefit(benefits: Benefit[]): BenefitMeasurementIndicator {
-    this.benefit = benefits.filter((b) => b.id === this.benefitId && b.siteId === this.siteId)[0]
+    this.Benefit = benefits.filter(
+      (b) => b.Id === this.BenefitItemId && b.SiteId === this.SiteId
+    )[0]
     return this
   }
 }