+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/ProjectProperties/ProjectProperty/index.tsx
@@ -1,10 +1,9 @@
 import { DisplayMode } from '@microsoft/sp-core-library'
-import { getId } from '@uifabric/utilities'
 import { Toggle } from 'office-ui-fabric-react/lib/Toggle'
 import * as strings from 'ProjectWebPartsStrings'
 import * as React from 'react'
-import { IProjectPropertyProps } from './types'
 import styles from './ProjectProperty.module.scss'
+import { IProjectPropertyProps } from './types'
 
 export const ProjectProperty = ({
   model,
@@ -13,13 +12,11 @@ export const ProjectProperty = ({
   onFieldExternalChanged,
   showFieldExternal
 }: IProjectPropertyProps) => {
-  const id = getId(`projectproperty_${model.internalName}`.toLowerCase())
-
   switch (displayMode) {
     case DisplayMode.Edit: {
       const defaultChecked = showFieldExternal ? showFieldExternal[model.internalName] : false
       return (
-        <div id={id} className={styles.projectProperty} title={model.description} style={style}>
+        <div className={styles.projectProperty} title={model.description} style={style}>
           <div className={styles.label}>{model.displayName}</div>
           <div className={styles.value}>
             <Toggle
@@ -34,7 +31,7 @@ export const ProjectProperty = ({
     }
     default: {
       return (
-        <div id={id} className={styles.projectProperty} title={model.description} style={style}>
+        <div className={styles.projectProperty} title={model.description} style={style}>
           <div className={styles.label}>{model.displayName}</div>
           <div className={styles.value}>{model.value}</div>
         </div>