+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/index.tsx
@@ -1,16 +1,15 @@
 import { DisplayMode } from '@microsoft/sp-core-library'
-import { stringIsNullOrEmpty, TypedHash } from '@pnp/common'
+import { stringIsNullOrEmpty } from '@pnp/common'
 import { LogLevel } from '@pnp/logging'
-import { WebPartTitle } from '@pnp/spfx-controls-react/lib/WebPartTitle'
 import { MessageBarType } from 'office-ui-fabric-react/lib/MessageBar'
 import { IProgressIndicatorProps } from 'office-ui-fabric-react/lib/ProgressIndicator'
-import { Spinner } from 'office-ui-fabric-react/lib/Spinner'
 import { format } from 'office-ui-fabric-react/lib/Utilities'
 import { PortalDataService } from 'pp365-shared/lib/services'
 import { parseUrlHash, sleep } from 'pp365-shared/lib/util'
 import * as strings from 'ProjectWebPartsStrings'
 import { ConfirmAction, ConfirmDialog } from 'pzl-spfx-components/lib/components/ConfirmDialog'
-import React from 'react'
+import React, { Fragment } from 'react'
+import { isEmpty } from 'underscore'
 import SPDataAdapter from '../../data'
 import { BaseWebPartComponent } from '../BaseWebPartComponent'
 import { ProgressDialog } from '../ProgressDialog'
@@ -19,7 +18,6 @@ import { Actions } from './Actions'
 import styles from './ProjectInformation.module.scss'
 import { ProjectProperties } from './ProjectProperties'
 import { ProjectPropertyModel } from './ProjectProperties/ProjectProperty'
-import { StatusReports } from './StatusReports'
 import {
   IProjectInformationData,
   IProjectInformationProps,
@@ -32,21 +30,15 @@ export class ProjectInformation extends BaseWebPartComponent<
   IProjectInformationState
 > {
   public static defaultProps: Partial<IProjectInformationProps> = {
-    statusReportsCount: 0,
     page: 'Frontpage'
   }
   private _portalDataService: PortalDataService
 
-  /**
-   * Constructor
-   *
-   * @param {IProjectInformationProps} props Props
-   */
   constructor(props: IProjectInformationProps) {
     super('ProjectInformation', props, { loading: true })
     this._portalDataService = new PortalDataService().configure({
-      urlOrWeb: props.hubSite.web,
-      siteId: props.siteId
+      urlOrWeb: this.props.hubSite.web,
+      siteId: this.props.siteId
     })
   }
 
@@ -62,21 +54,16 @@ export class ProjectInformation extends BaseWebPartComponent<
   }
 
   public render() {
-    if (this.state.hidden) {
-      return null
-    }
+    if (this.state.hidden) return null
     return (
-      <div className={styles.projectInformation}>
+      <div className={styles.root}>
         <div className={styles.container}>
-          <div hidden={this.props.displayMode === DisplayMode.Edit}>
-            <WebPartTitle
-              displayMode={DisplayMode.Read}
-              title={this.props.title}
-              // eslint-disable-next-line @typescript-eslint/no-empty-function
-              updateProperty={() => {}}
-            />
+          <div className={styles.header}>
+            <span role='heading' aria-level={2}>
+              {this.props.title}
+            </span>
           </div>
-          {this._contents}
+          {this.getContent()}
         </div>
       </div>
     )
@@ -85,12 +72,8 @@ export class ProjectInformation extends BaseWebPartComponent<
   /**
    * Contents
    */
-  private get _contents() {
-    if (this.state.loading) {
-      return !stringIsNullOrEmpty(this.props.title) ? (
-        <Spinner label={format(strings.LoadingText, this.props.title.toLowerCase())} />
-      ) : null
-    }
+  private getContent() {
+    if (this.state.loading) return null
     if (this.state.error) {
       return (
         <UserMessage
@@ -101,10 +84,10 @@ export class ProjectInformation extends BaseWebPartComponent<
       )
     }
 
-    const { statusReports, editFormUrl, versionHistoryUrl } = this.state.data
+    const { editFormUrl, versionHistoryUrl } = this.state.data
 
     return (
-      <>
+      <Fragment>
         <ProjectProperties
           title={this.props.title}
           properties={this.state.properties}
@@ -113,12 +96,6 @@ export class ProjectInformation extends BaseWebPartComponent<
           onFieldExternalChanged={this.props.onFieldExternalChanged}
           showFieldExternal={this.props.showFieldExternal}
           propertiesList={!stringIsNullOrEmpty(this.state.data.propertiesListId)}
-        />
-        <StatusReports
-          title={this.props.statusReportsHeader}
-          statusReports={statusReports}
-          urlSourceParam={document.location.href}
-          hidden={this.props.statusReportsCount === 0}
         />
         <UserMessage {...this.state.message} />
         <Actions
@@ -136,7 +113,7 @@ export class ProjectInformation extends BaseWebPartComponent<
         />
         <ProgressDialog {...this.state.progress} />
         {this.state.confirmActionProps && <ConfirmDialog {...this.state.confirmActionProps} />}
-      </>
+      </Fragment>
     )
   }
 
@@ -237,28 +214,24 @@ export class ProjectInformation extends BaseWebPartComponent<
   /**
    * Transform properties from entity item and configuration
    *
-   * @param {TypedHash} fieldValuesText Field values as text
    * @param {IProjectInformationData} data Data
    */
-  private _transformProperties(fieldValuesText: TypedHash<any>, data: IProjectInformationData) {
+  private _transformProperties({ columns, fields, fieldValuesText }: IProjectInformationData) {
     const fieldNames: string[] = Object.keys(fieldValuesText).filter((fieldName) => {
-      const [field] = data.fields.filter((fld) => fld.InternalName === fieldName)
-      if (!field) {
-        return false
-      }
+      const [field] = fields.filter((fld) => fld.InternalName === fieldName)
+      if (!field) return false
       if (
-        data.columns.length === 0 &&
+        isEmpty(columns) &&
         ((this.props.showFieldExternal || {})[fieldName] || this.props.skipSyncToHub)
       ) {
         return true
       }
-      const [column] = data.columns.filter((c) => c.internalName === fieldName)
-      if (column) return column.isVisible(this.props.page)
-      return false
+      const [column] = columns.filter((c) => c.internalName === fieldName)
+      return column ? column.isVisible(this.props.page) : false
     })
-    const properties = fieldNames.map((fieldName) => {
-      const [field] = data.fields.filter((fld) => fld.InternalName === fieldName)
-      return new ProjectPropertyModel(field, fieldValuesText[fieldName])
+    const properties = fieldNames.map((fn) => {
+      const [field] = fields.filter((fld) => fld.InternalName === fn)
+      return new ProjectPropertyModel(field, fieldValuesText[fn])
     })
     return properties
   }
@@ -268,38 +241,21 @@ export class ProjectInformation extends BaseWebPartComponent<
    */
   private async _fetchData(): Promise<Partial<IProjectInformationState>> {
     try {
-      if (!SPDataAdapter.isConfigured) {
-        SPDataAdapter.configure(this.context, {
-          siteId: this.props.siteId,
-          webUrl: this.props.webUrl,
-          hubSiteUrl: this.props.hubSite.url,
-          logLevel: sessionStorage.DEBUG || DEBUG ? LogLevel.Info : LogLevel.Warning
-        })
-      }
-
-      const [columnConfig, propertiesData] = await Promise.all([
+      SPDataAdapter.configure(this.props.webPartContext, {
+        siteId: this.props.siteId,
+        webUrl: this.props.webUrl,
+        hubSiteUrl: this.props.hubSite.url,
+        logLevel: sessionStorage.DEBUG || DEBUG ? LogLevel.Info : LogLevel.Warning
+      })
+      const [columns, propertiesData] = await Promise.all([
         this._portalDataService.getProjectColumns(),
         SPDataAdapter.project.getPropertiesData()
       ])
-
       const data: IProjectInformationData = {
-        statusReports: [],
-        columns: columnConfig,
+        columns,
         ...propertiesData
       }
-
-      if (this.props.statusReportsCount > 0) {
-        data.statusReports = await this._portalDataService.getStatusReports({
-          top: this.props.statusReportsCount
-        })
-      }
-
-      const properties = this._transformProperties(data.fieldValuesText, data)
-
-      this.logInfo('Succesfully retrieved data.', '_fetchData', {
-        propertiesListId: data.propertiesListId
-      })
-
+      const properties = this._transformProperties(data)
       return { data, properties }
     } catch (error) {
       this.logError('Failed to retrieve data.', '_fetchData', error)