+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Every change is marked with Pull Request ID.
 - Issues with custom project fields #378
 - Fix for visible check in project information web part #385
 - Setting phase check list and planner task list as visible #389
+- Flexible portfolio aggregation web part #395
 - Copying for documents to new projects #399
 
 ## 1.2.4 - 30.11.2020